+++ CHANGELOG.md
@@ -13,7 +13,7 @@ it in future.
 ## Unreleased
 ### Added
 
-* Added GenieOption to read simulations with the GENIE Geometry Driver (i.e. gevgen_fnal) 
+* Added GenieOption to read simulations with the GENIE Geometry Driver (i.e. gevgen_fnal)
 
 ### Changed